update error analyzis script
