+++ iyal_quality_analyzer/error_analyzis/analyze_gnanam_legacy.py
@@ -1,11 +1,30 @@
 import os
 import sys
 from docx import Document
+import csv
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 from iyal_quality_analyzer.inference_base.inference import Inference
 from iyal_quality_analyzer.quality_analyzer import single_sentence_quality_analyzer
+from iyal_quality_analyzer.utils.legacy_converter.legacy_converter import convert_legacy_to_unicode
 
+def update_csv(input_word: str, input_type: str, output: str, actual_output: str, csv_file: str = 'E:\___MORA\FYP\FinalRepos\Project-IYAL\error_analyzis\output.csv'):
+    """
+    Updates the CSV file with the given data.
+
+    Args:
+        input_word (str): The input word.
+        input_type (str): The input type.
+        output (str): The output.
+        actual_output (str): The actual output.
+        csv_file (str): The path to the CSV file.
+    """
+    file_exists = os.path.isfile(csv_file)
+    with open(csv_file, mode='a', newline='', encoding='utf-8') as file:
+        writer = csv.writer(file, quoting=csv.QUOTE_ALL)
+        if not file_exists:
+            writer.writerow(['inputWord', 'inputType', 'output', 'actualOutput'])
+        writer.writerow([input_word, input_type, output, actual_output])
 
 def analyze_text_files(directory: str, model: Inference, encoding: str = "bamini2utf8"):
     """
@@ -16,7 +35,6 @@ def analyze_text_files(directory: str, model: Inference, encoding: str = "bamini
         model (Inference): The model to use for legacy font classification.
         encoding (str): The encoding of the input text (e.g., bamini2utf8, etc.).
     """
-    breakpoint()
     for filename in os.listdir(directory):
         if filename.endswith(".TXT"):
             file_path = os.path.join(directory, filename)
@@ -26,6 +44,14 @@ def analyze_text_files(directory: str, model: Inference, encoding: str = "bamini
                         try:
                             output, results = single_sentence_quality_analyzer(
                                 model, line.strip(), [], encoding)
+                                # Calculate actual output
+
+                            # Update CSV file
+                            for result in results:
+                                input_word = result["inputWord"]
+                                actual_output = convert_legacy_to_unicode(input_word, 'bamini2utf8')
+                                update_csv(input_word, result["inputType"], result["output"], actual_output)
+                                    
                         except Exception as e:
                             print(f"Error processing line: {str(e)}")
                             continue
@@ -38,6 +64,11 @@ def analyze_text_files(directory: str, model: Inference, encoding: str = "bamini
                     try:
                         output, results = single_sentence_quality_analyzer(
                             model, line, [], encoding)
+                        
+                        for result in results:
+                            input_word = result["inputWord"]
+                            actual_output = convert_legacy_to_unicode(input_word, 'bamini2utf8')
+                            update_csv(input_word, result["inputType"], result["output"], actual_output)
                     except Exception as e:
                         print(f"Error processing line: {str(e)}")
                         continue
